mixradio

+++ backends/sql/sqlite/database.class.php
@@ -339,7 +339,7 @@ class database extends data_base {
 	}
 
 	public function create_toptracks_table() {
-		$name = lastfm_radio::get_seed_table_name();
+		$name = everywhere_radio::get_seed_table_name();
 		if ($this->generic_sql_query("CREATE TABLE IF NOT EXISTS ".$name."(".
 			"topindex INTEGER PRIMARY KEY NOT NULL UNIQUE, ".
 			"Type INTEGER NOT NULL, ".
@@ -356,7 +356,7 @@ class database extends data_base {
 	}
 
 	public function add_toptrack($type, $artist, $title) {
-		$name = lastfm_radio::get_seed_table_name();
+		$name = everywhere_radio::get_seed_table_name();
 		$this->sql_prepare_query(true, null, null, null,
 			"INSERT OR IGNORE INTO ".$name." (Type, Artist, Title) VALUES (?, ? ,?)",
 			$type,
@@ -366,17 +366,31 @@ class database extends data_base {
 	}
 
 	public function create_radio_uri_table() {
-		$name = lastfm_radio::get_uri_table_name();
+		$name = everywhere_radio::get_uri_table_name();
 		if ($this->generic_sql_query("CREATE TABLE IF NOT EXISTS ".$name."(".
 			"uriindex INTEGER PRIMARY KEY NOT NULL UNIQUE, ".
+			"Artist VARCHAR(100) COLLATE NOCASE, ".
+			"Title VARCHAR(255) COLLATE NOCASE, ".
 			"Uri TEXT)", true))
 		{
 			logger::log("SQLITE",$name,"OK");
+			$this->generic_sql_query("DELETE FROM ".$name);
+			$this->generic_sql_query("CREATE UNIQUE INDEX IF NOT EXISTS nodupes_".$name." ON ".$name." (Artist, Title)");
 		} else {
 			$err = $this->mysqlc->errorInfo()[2];
 			return array(false, "Error While Checking ".$name." : ".$err);
 		}
-		$this->generic_sql_query("DELETE FROM ".$name);
 	}
+
+	public function add_smart_uri($uri, $artist, $title) {
+		$name = everywhere_radio::get_uri_table_name();
+		$this->sql_prepare_query(true, null, null, null,
+			"INSERT OR IGNORE INTO ".$name." (Artist, Title, Uri) VALUES (?, ? ,?)",
+			$artist,
+			$title,
+			$uri
+		);
+	}
+
 }
 ?>
